requirements/prod.txt
django==5.0.7
python-dotenv==1.0.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
mysqlclient==2.2.4
requests==2.32.3
django-ipware==7.0.1